Log a xml

+++ src/pluginPackage/NewActionsForPlayButton.java
@@ -17,7 +17,21 @@ import org.eclipse.jgit.lib.Repository;
 import org.eclipse.jgit.transport.RefSpec;
 import org.eclipse.jgit.transport.UsernamePasswordCredentialsProvider;
 import org.jetbrains.annotations.NotNull;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.Node;
+import org.w3c.dom.NodeList;
+import org.xml.sax.SAXException;
 
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.transform.OutputKeys;
+import javax.xml.transform.Transformer;
+import javax.xml.transform.TransformerException;
+import javax.xml.transform.TransformerFactory;
+import javax.xml.transform.dom.DOMSource;
+import javax.xml.transform.stream.StreamResult;
 import java.io.*;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
@@ -129,28 +143,63 @@ public class NewActionsForPlayButton extends AbstractProjectComponent {
 
                 try {
                     Thread.sleep(3000);
-                    File log = new File(workspaceFile.logPath);
                     BufferedReader br = null;
                     FileReader fr = null;
+                    FileWriter fileWriter;
+                    PrintWriter printWriter = null;
 
-                    //Punteros para escribir en archivo log.log, para no sobreescribir se pone append: true
-                    FileWriter fileWriter = new FileWriter(workspaceFile.logPath, true);
-                    PrintWriter printWriter = new PrintWriter(fileWriter);
-                    printWriter.print("<log_project id ='"+ copyFiles.df.format(dateobj) +"' date ='"
-                            + df.format(dateobj) + "' name ='" + project.getName() + "'>\n");
-                    fr = new FileReader(project.getBasePath() + workspaceFile.consoleOutputPath);
-                    br = new BufferedReader(fr);
+                    if (!FileUtils.fileExists(workspaceFile.logPath)){
+                        File log = new File(workspaceFile.logPath);
+                        log.createNewFile();
 
-                    String sCurrentLine;
-
-                    //Copiando contenido de output.txt a log.log.
-                    while ((sCurrentLine = br.readLine()) != null) {
-                        printWriter.print(sCurrentLine + "\n");
+                        fileWriter = new FileWriter(workspaceFile.logPath, true);
+                        printWriter = new PrintWriter(fileWriter);
+                        printWriter.print("<?xml version=\"1.0\" encoding=\"UTF-8\"?>" + "\n");
+                        printWriter.print("<output>" + "\n");
+                        printWriter.print("</output>");
+                        printWriter.close();
                     }
-                    printWriter.print("</log_project>\n");
-                    printWriter.close();
 
-                } catch (IOException | InterruptedException ex) {
+
+                    try{
+                        DocumentBuilderFactory dbFactory = DocumentBuilderFactory.newInstance();
+                        DocumentBuilder dBuilder = dbFactory.newDocumentBuilder();
+                        Document doc = dBuilder.parse(workspaceFile.logPath);
+                        Element root  = doc.getDocumentElement();
+                        Element eNewComponent = doc.createElement("log_project");
+                        eNewComponent.setAttribute("id", "" + copyFiles.df.format(dateobj));
+                        eNewComponent.setAttribute("date", "" + df.format(dateobj));
+                        eNewComponent.setAttribute("name_project", "" + project.getName());
+
+                        fr = new FileReader(project.getBasePath() + workspaceFile.consoleOutputPath);
+                        br = new BufferedReader(fr);
+
+                        String sCurrentLine;
+                        //Copiando contenido de output.txt a log.log.
+                        while ((sCurrentLine = br.readLine()) != null) {
+                            eNewComponent.appendChild(doc.createTextNode("" + sCurrentLine + "\n"));
+                        }
+                        root.appendChild(eNewComponent);
+                        TransformerFactory transformerFactory = TransformerFactory.newInstance();
+                        Transformer transformer = transformerFactory.newTransformer();
+                        transformer.setOutputProperty(OutputKeys.ENCODING, "UTF-8");
+                        transformer.setOutputProperty(OutputKeys.INDENT, "yes");
+                        DOMSource source = new DOMSource(doc);
+                        StreamResult result = new StreamResult(new File(workspaceFile.logPath));
+                        transformer.transform(source, result);
+                    } catch (ParserConfigurationException ex){
+                        ex.printStackTrace();
+                    } catch(SAXException sx){
+                        sx.printStackTrace();
+                    } catch(TransformerException ax) {
+                        ax.printStackTrace();
+                    } finally {
+                        fr.close();
+                        br.close();
+                    }
+
+                } catch (IOException | InterruptedException | NullPointerException ex) {
+                    ex.printStackTrace();
                     System.out.println("Error con manejo de archivos.");
                 }
